Split out pairlist_commands
