+++ freqtrade/commands/__init__.py
@@ -13,8 +13,8 @@ from freqtrade.commands.list_commands import (start_list_exchanges,
                                               start_list_timeframes)
 from freqtrade.commands.optimize_commands import (start_backtesting,
                                                   start_edge, start_hyperopt)
+from freqtrade.commands.pairlist_commands import start_test_pairlist
 from freqtrade.commands.plot_commands import (start_plot_dataframe,
                                               start_plot_profit)
 from freqtrade.commands.trade_commands import start_trading
-from freqtrade.commands.utils import (setup_utils_configuration,
-                                      start_test_pairlist)
+from freqtrade.commands.utils import setup_utils_configuration
